feat(secret-provider): Add AWS Secret Provider (#5291)
* feat(secret-provider): Add AWS Secret Provider

* feat(aws-secret-provider): Make separate package for aws

* feat(aws-secret-provider): Move META-INF files to corret location and add tests

* feat(aws-secret-provider): Add aws secret provider

(cherry picked from commit db0253617baf7b0e8b1d7637b9e7c1093d7f96e1)
