--- secret-providers/gcp-secret-provider/src/main/resources/META-INF/services/io.camunda.connector.api.secret.SecretProvider
@@ -1,1 +1,0 @@
-io.camunda.connector.runtime.cloud.GcpSecretManagerSecretProvider
