+++ secret-providers/gcp/src/main/resources/META-INF/services/io.camunda.connector.api.secret.SecretProvider
@@ -1,0 +1,1 @@
+io.camunda.connector.secret.providers.GcpSecretProvider
