+++ secret-providers/pom.xml
@@ -1,0 +1,27 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <modelVersion>4.0.0</modelVersion>
+
+  <parent>
+    <groupId>io.camunda.connector</groupId>
+    <artifactId>connector-parent</artifactId>
+    <relativePath>../parent/pom.xml</relativePath>
+    <version>8.6.19-rc2</version>
+  </parent>
+
+  <artifactId>secret-providers</artifactId>
+  <packaging>pom</packaging>
+
+  <properties>
+    <version.google-libraries-bom>26.68.0</version.google-libraries-bom>
+    <maven.compiler.source>21</maven.compiler.source>
+    <maven.compiler.target>21</maven.compiler.target>
+    <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+  </properties>
+
+  <name>Connector Secret Provider</name>
+  <modules>
+    <module>secret-providers-base</module>
+    <module>gcp</module>
+    <module>aws</module>
+  </modules>
+</project>
